+++ server/controllers/admin/Owner.controller.js
@@ -1,6 +1,7 @@
 const Owner = require("../../models/Owner.model");
 const deleteFile = require("../../utils/image.cleaner");
 const path = require("path");
+const validationResult = require("express-validator").validationResult;
 
 /**
  * title:createOwner
@@ -8,29 +9,40 @@ const path = require("path");
  * Method:POST
  */
 module.exports.createOwner = (req, res, next) => {
-  let owner = new Owner();
-  owner.name = req.body.name;
-  owner.about = req.body.about;
-  if (req.file) owner.photo = req.file.path;
+  // check incomming data is valid?
+  let errors = validationResult(req).array();
+  if (errors.length > 0) {
+    let response = {
+      status: "Validation Errors",
+      errors: errors,
+    };
+    res.status(403).json(response);
+  } else {
+    //data is valid.
+    let owner = new Owner();
+    owner.name = req.body.name;
+    owner.about = req.body.about;
+    if (req.file) owner.photo = req.file.path;
 
-  owner
-    .save()
-    .then((_owner) => {
-      let response = {
-        status: "Success",
-        msg: "new Owner created successfully.",
-        owner: _owner,
-      };
-      res.status(201).json(response);
-    })
-    .catch((err) => {
-      let response = {
-        status: "Error",
-        msg: err.message,
-        error: err,
-      };
-      res.status(500).json(response);
-    });
+    owner
+      .save()
+      .then((_owner) => {
+        let response = {
+          status: "Success",
+          msg: "new Owner created successfully.",
+          owner: _owner,
+        };
+        res.status(201).json(response);
+      })
+      .catch((err) => {
+        let response = {
+          status: "Error",
+          msg: err.message,
+          error: err,
+        };
+        res.status(500).json(response);
+      });
+  }
 };
 
 /**
@@ -64,36 +76,47 @@ module.exports.getOwners = (req, res, next) => {
  * Method:PUT
  */
 module.exports.updateOwner = (req, res, next) => {
-  let { id } = req.body;
-  Owner.findById(id)
-    .then((owner) => {
-      owner.name = req.body.name;
-      owner.about = req.body.about;
+  // check incomming data is valid?
+  let errors = validationResult(req).array();
+  if (errors.length > 0) {
+    let response = {
+      status: "Validation Errors",
+      errors: errors,
+    };
+    res.status(403).json(response);
+  } else {
+    //data is valid.
+    let { id } = req.body;
+    Owner.findById(id)
+      .then((owner) => {
+        owner.name = req.body.name;
+        owner.about = req.body.about;
 
-      if (req.file) {
-        if (owner.photo) {
-          deleteFile(path.join(__dirname, "..", "..", owner.photo));
+        if (req.file) {
+          if (owner.photo) {
+            deleteFile(path.join(__dirname, "..", "..", owner.photo));
+          }
+          owner.photo = req.file.path;
         }
-        owner.photo = req.file.path;
-      }
 
-      owner.save().then((_owner) => {
+        owner.save().then((_owner) => {
+          let response = {
+            status: "Success",
+            msg: "Owner updated successfully.",
+            owner: _owner,
+          };
+          res.status(201).json(response);
+        });
+      })
+      .catch((err) => {
         let response = {
-          status: "Success",
-          msg: "Owner updated successfully.",
-          owner: _owner,
+          status: "Error",
+          msg: err.message,
+          error: err,
         };
-        res.status(201).json(response);
+        res.status(500).json(response);
       });
-    })
-    .catch((err) => {
-      let response = {
-        status: "Error",
-        msg: err.message,
-        error: err,
-      };
-      res.status(500).json(response);
-    });
+  }
 };
 
 /**
@@ -102,22 +125,33 @@ module.exports.updateOwner = (req, res, next) => {
  * Method:DELETE
  */
 module.exports.deleteOwner = (req, res, next) => {
-  let { id } = req.body;
-  Owner.findByIdAndRemove(id)
-    .then((_owner) => {
-      let response = {
-        status: "Success",
-        msg: "Owner deleted successfully.",
-        owner: _owner,
-      };
-      res.status(200).json(response);
-    })
-    .catch((err) => {
-      let response = {
-        status: "Error",
-        msg: err.message,
-        error: err,
-      };
-      res.status(500).json(response);
-    });
+  // check incomming data is valid?
+  let errors = validationResult(req).array();
+  if (errors.length > 0) {
+    let response = {
+      status: "Validation Errors",
+      errors: errors,
+    };
+    res.status(403).json(response);
+  } else {
+    //data is valid.
+    let { id } = req.body;
+    Owner.findByIdAndRemove(id)
+      .then((_owner) => {
+        let response = {
+          status: "Success",
+          msg: "Owner deleted successfully.",
+          owner: _owner,
+        };
+        res.status(200).json(response);
+      })
+      .catch((err) => {
+        let response = {
+          status: "Error",
+          msg: err.message,
+          error: err,
+        };
+        res.status(500).json(response);
+      });
+  }
 };
